update gumtree output
